refactor: move InputManager and DiscreteActionHandler outside of handlers package (remove it fully)
