+++ src/main/java/yesman/epicfight/client/input/InputUtils.java
@@ -5,7 +5,7 @@ import net.minecraft.client.player.Input;
 import net.minecraft.client.player.LocalPlayer;
 import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.NotNull;
-import yesman.epicfight.api.client.input.handlers.InputManager;
+import yesman.epicfight.api.client.input.InputManager;
 
 /// Internal utility for simplified input checks.
 /// Consumers should avoid calling these methods, as they may break in future versions.
